Patch: changed create_ipes to work with mock catalog (#75)
* initial commit: changed create_ipes to work with mock catalog

* fix: changed notebook to comply with new param functions
